requirements.txt
six>1.11
numpy>=1.15
wxpython>=4.02
wxmplot>=0.9.38
pyepics>=3.4.0
sqlalchemy>=
matplotlib>=3.0
sqlalchemy>=0.9
lmfit>=0.9.13
pyshortcuts>=1.5
pyepics>=3.4.0
wxpython>=4.0.3
wxutils>=0.2.3
wxmplot>=0.9.37
yaml>=5.0
xraydb
pip
pillow
pyfai
